review fix part1
