+++ internal/server/handlers/add_file.go
@@ -9,10 +9,11 @@ import (
 )
 
 // AddItemFileHandler обработчик роута: POST /api/items/file
-func AddItemFileHandler(s storage.Storage) http.HandlerFunc {
+func AddItemFileHandler(s storage.Storage, maxUploadFileSize int64) http.HandlerFunc {
 	return func(res http.ResponseWriter, req *http.Request) {
 		var input storage.CommonData
 		var fileBytes []byte
+		req.Body = http.MaxBytesReader(res, req.Body, maxUploadFileSize)
 		file, handler, err := req.FormFile("file")
 		if err != nil {
 			http.Error(res, err.Error(), http.StatusBadRequest)
